feat: builds for wasm
Adds editor configuration, a makefile, documentation, and some
code adjustents to provide bindings to the `OpfsBackend` on WASM.

In general, we wouldn't expect web users to construct this directly:
it doesn't (yet) give them anything that they can't more conveniently
get by using the relevant APIs directly.

However, it's necessary to run this on a web worker, because
the necessary trait is synchronous. This means that we need to support
two use cases: where the "main" part of the rust code also runs on
a web worker, and where it does not.

Where the "main" part of the rust code runs on a web worker, then
things are easy: just add this as a dependency as shown in the readme
and everything stays fast and easy on the Rust side of the FFI boundary.

Where the "main" part of the rust code runs on the main thread, however,
then things are less easy. We need to let JS people launch this part
of the code on the worker, and provide a handle which they can give to
OPFS on the main thread. That handle will also implement the
`StorageBackend` trait, but it will do so by communicating with the
web worker. This part is not yet built. Still, having the design
in mind means that we know we will need a standalone struct which can
be created in JS-land, which we can then call `.handle()` on to
get that handle to pass back to the main thread. So that's why we have
this thing. We just need to write the handle.