wip: set up web worker demo to run in the browser
Turns out that Bun doesn't support OPFS, just failing at runtime
when you try to do that.

It also turns out that as far as Rust's stdlib `Path::parent` is
concerned, the parent of "opfs" is `Some("")`, which is less helpful
than one would hope.

So first we fix the bit where we try to open an empty-named directory,
and then we set things up so the demo runs in the browser.